functions/requirements.txt
firebase_functions~=0.4.2
firebase_admin~=6.8.0
python-dotenv~=1.1.0